Cut and paste DefaultEffectEditor into new files
